upgrade GEOS to 3.10.0

+++ deps/libgeos/geos/include/geos/geomgraph/Depth.h
@@ -43,14 +43,14 @@ class GEOS_DLL Depth {
 public:
     static int depthAtLocation(geom::Location location);
     Depth();
-    virtual ~Depth() = default; // FIXME: shoudn't be virtual!
+    virtual ~Depth() = default; // FIXME: shouldn't be virtual!
     int getDepth(int geomIndex, int posIndex) const;
     void setDepth(int geomIndex, int posIndex, int depthValue);
     geom::Location getLocation(int geomIndex, int posIndex) const;
     void add(int geomIndex, int posIndex, geom::Location location);
     bool isNull() const;
-    bool isNull(int geomIndex) const;
-    bool isNull(int geomIndex, int posIndex) const;
+    bool isNull(uint8_t geomIndex) const;
+    bool isNull(uint8_t geomIndex, uint8_t posIndex) const;
     int getDelta(int geomIndex) const;
     void normalize();
     void add(const Label& lbl);
